# account_manager/models/change_password_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from account_manager.models.base_model_ import Model
from account_manager import util


class ChangePasswordRequest(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, old_password=None, new_password=None, new_password_repeat=None):  # noqa: E501
        """ChangePasswordRequest - a model defined in OpenAPI

        :param old_password: The old_password of this ChangePasswordRequest.  # noqa: E501
        :type old_password: str
        :param new_password: The new_password of this ChangePasswordRequest.  # noqa: E501
        :type new_password: str
        :param new_password_repeat: The new_password_repeat of this ChangePasswordRequest.  # noqa: E501
        :type new_password_repeat: str
        """
        self.openapi_types = {
            'old_password': str,
            'new_password': str,
            'new_password_repeat': str
        }

        self.attribute_map = {
            'old_password': 'old_password',
            'new_password': 'new_password',
            'new_password_repeat': 'new_password_repeat'
        }

        self._old_password = old_password
        self._new_password = new_password
        self._new_password_repeat = new_password_repeat

    @classmethod
    def from_dict(cls, dikt) -> 'ChangePasswordRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ChangePasswordRequest of this ChangePasswordRequest.  # noqa: E501
        :rtype: ChangePasswordRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def old_password(self):
        """Gets the old_password of this ChangePasswordRequest.


        :return: The old_password of this ChangePasswordRequest.
        :rtype: str
        """
        return self._old_password

    @old_password.setter
    def old_password(self, old_password):
        """Sets the old_password of this ChangePasswordRequest.


        :param old_password: The old_password of this ChangePasswordRequest.
        :type old_password: str
        """
        if old_password is None:
            raise ValueError("Invalid value for `old_password`, must not be `None`")  # noqa: E501
        if old_password is not None and len(old_password) > 30:
            raise ValueError("Invalid value for `old_password`, length must be less than or equal to `30`")  # noqa: E501
        if old_password is not None and len(old_password) < 6:
            raise ValueError("Invalid value for `old_password`, length must be greater than or equal to `6`")  # noqa: E501

        self._old_password = old_password

    @property
    def new_password(self):
        """Gets the new_password of this ChangePasswordRequest.


        :return: The new_password of this ChangePasswordRequest.
        :rtype: str
        """
        return self._new_password

    @new_password.setter
    def new_password(self, new_password):
        """Sets the new_password of this ChangePasswordRequest.


        :param new_password: The new_password of this ChangePasswordRequest.
        :type new_password: str
        """
        if new_password is None:
            raise ValueError("Invalid value for `new_password`, must not be `None`")  # noqa: E501
        if new_password is not None and len(new_password) > 30:
            raise ValueError("Invalid value for `new_password`, length must be less than or equal to `30`")  # noqa: E501
        if new_password is not None and len(new_password) < 6:
            raise ValueError("Invalid value for `new_password`, length must be greater than or equal to `6`")  # noqa: E501

        self._new_password = new_password

    @property
    def new_password_repeat(self):
        """Gets the new_password_repeat of this ChangePasswordRequest.


        :return: The new_password_repeat of this ChangePasswordRequest.
        :rtype: str
        """
        return self._new_password_repeat

    @new_password_repeat.setter
    def new_password_repeat(self, new_password_repeat):
        """Sets the new_password_repeat of this ChangePasswordRequest.


        :param new_password_repeat: The new_password_repeat of this ChangePasswordRequest.
        :type new_password_repeat: str
        """
        if new_password_repeat is None:
            raise ValueError("Invalid value for `new_password_repeat`, must not be `None`")  # noqa: E501
        if new_password_repeat is not None and len(new_password_repeat) > 30:
            raise ValueError("Invalid value for `new_password_repeat`, length must be less than or equal to `30`")  # noqa: E501
        if new_password_repeat is not None and len(new_password_repeat) < 6:
            raise ValueError("Invalid value for `new_password_repeat`, length must be greater than or equal to `6`")  # noqa: E501

        self._new_password_repeat = new_password_repeat