# account_manager/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: vasco.m.campos@inesctec.pt
    name: Vasco Campos
  description: 'User Account Manager Service OpenAPI definition. This service has
    the following functions: register, login, change user credentials and update it''s
    profile settings. Find out more: [User Account Manager Service documentation](https://gitlab.inesctec.pt/cpes/european-projects/interconnect/hems/hems-documentation/-/blob/master/Microservices/User-Account-Manager-Service.adoc)'
  title: User Account Manager Service
  version: 1.2.0
servers:
- url: https://interconnect-dev.inesctec.pt/api/account
security:
- {}
- Bearer: []
...
paths:
...
  /change-password:
    description: Change password
    post:
      operationId: change_password_post
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: true
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        "200":
          description: Successfully changed password.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Change password
      tags:
      - Password management
      x-openapi-router-controller: account_manager.controllers.password_management_controller
...
  /forgot-password:
    description: Request to sent a link to the email
    post:
      operationId: forgot_password_post
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
        required: true
      responses:
        "200":
          description: Successfully requested password change.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Forgot password
      tags:
      - Password management
      x-openapi-router-controller: account_manager.controllers.password_management_controller
  /forgot-password/{token}:
    description: Change password after clicking the link sent to the user email.
    get:
      operationId: forgot_password_token_get
      parameters:
      - explode: false
        in: path
        name: token
        required: true
        schema:
          $ref: '#/components/schemas/ConfirmationToken'
        style: simple
      responses:
        "200":
          description: Web page with button to send the POST request to set new password.
        "404":
          description: Token not found or expired - Web page with status 404.
      summary: Endpoint to get the webpage with the form to reset password by clicking
        the link sent to the e-mail.
      tags:
      - Password management
      x-openapi-router-controller: account_manager.controllers.password_management_controller
    post:
      operationId: forgot_password_token_post
      parameters:
      - explode: false
        in: path
        name: token
        required: true
        schema:
          $ref: '#/components/schemas/ConfirmationToken'
        style: simple
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ForgotPasswordNewPassword'
        required: true
      responses:
        "200":
          description: Account confirmed - Web page with status 200.
        "400":
          description: Bad request - Web page with status 400.
        "404":
          description: Token not found or expired - Web page with status 404.
      summary: Endpoint to set the new password. This endpoint will be used by the
        webpage to reset the password (get endpoint of this request).
      tags:
      - Password management
      x-openapi-router-controller: account_manager.controllers.password_management_controller
...
components:
  headers:
    CorrelationId:
      explode: false
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'
      style: simple
...
  schemas:
    CorrelationId:
      format: uuid
      type: string
...
    ConfirmationToken:
      maxLength: 64
      minLength: 64
      type: string
    Authorization:
      format: byte
      type: string
...
    Password:
      example: Ja7ha6dUHs7knm
      format: password
      maxLength: 30
      minLength: 6
      type: string
...
    ChangePasswordRequest:
      description: Model containing the change password request
      example:
        old_password: Ja7ha6dUHs7knm
        new_password: Ja7ha6dUHs7knm
        new_password_repeat: Ja7ha6dUHs7knm
      properties:
        old_password:
          example: Ja7ha6dUHs7knm
          format: password
          maxLength: 30
          minLength: 6
          type: string
        new_password:
          example: Ja7ha6dUHs7knm
          format: password
          maxLength: 30
          minLength: 6
          type: string
        new_password_repeat:
          example: Ja7ha6dUHs7knm
          format: password
          maxLength: 30
          minLength: 6
          type: string
      required:
      - new_password
      - new_password_repeat
      - old_password
      type: object
    ForgotPasswordRequest:
      description: Model for the forgot password request body
      example:
        email: myemail@mycompany.com
      properties:
        email:
          example: myemail@mycompany.com
          format: email
          maxLength: 50
          title: email
          type: string
      required:
      - email
      type: object
    ForgotPasswordNewPassword:
      description: Model containing the new password to change after forgetting the
        old password
      properties:
        new_password:
          $ref: '#/components/schemas/Password'
        new_password_repeat:
          $ref: '#/components/schemas/Password'
      required:
      - new_password
      - new_password_repeat
      type: object
...
    Error:
      description: Error model
      properties:
        error:
          title: error
          type: string
      required:
      - error
      title: Error
      type: object
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: account_manager.controllers.security_controller_.info_from_Bearer